ruby側でfloatになっているものをswaggerではdoubleとして扱うように修正

+++ lib/ruby-swagger/grape/type.rb
@@ -91,7 +91,7 @@ module Swagger::Grape
         },
         'float' => {
           'type' => 'number',
-          'format' => 'float'
+          'format' => 'double'
         },
         'rack::multipart::uploadedfile' => {
           'type' => 'file'
